+++ src/test/java/io/github/torand/openapi2java/TestHelper.java
@@ -32,6 +32,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static io.github.torand.openapi2java.TestHelper.ConfigVariant.COMMON_HEADERS_FACTORY;
+import static io.github.torand.openapi2java.TestHelper.ConfigVariant.OIDC_CLIENT_ANNOTATION;
 import static io.github.torand.openapi2java.TestHelper.ConfigVariant.RESTEASY;
 import static io.github.torand.openapi2java.utils.StringUtils.removeLineBreaks;
 import static java.util.Collections.emptyList;
@@ -46,7 +47,8 @@ public final class TestHelper {
     public enum ConfigVariant {
         RESTEASY("Resteasy"),
         NO_HEADER_PARAM("NoHeaderParam"),
-        COMMON_HEADERS_FACTORY("CommonHeadersFactory");
+        COMMON_HEADERS_FACTORY("CommonHeadersFactory"),
+        OIDC_CLIENT_ANNOTATION("OidcClientAnnotation");
 
         String suffix;
 
@@ -97,6 +99,12 @@ public final class TestHelper {
         return baseOptions
             .withResourceNameSuffix(baseOptions.resourceNameSuffix() + "_" + COMMON_HEADERS_FACTORY.suffix)
             .withResourceClientHeadersFactoryOverride("io.github.torand.openapi2java.test.CommonClientHeadersFactory");
+    }
+
+    public static Options withOidcClientAnnotation(Options baseOptions) {
+        return baseOptions
+            .withResourceNameSuffix(baseOptions.resourceNameSuffix() + "_" + OIDC_CLIENT_ANNOTATION.suffix)
+            .withUseOidcClientAnnotation(true);
     }
 
     public static void assertSnippet(String path, String expectedSnippet) {